+++ src/main/java/codesquad/annotation/LoginCheck.java
@@ -1,0 +1,11 @@
+package codesquad.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+public @interface LoginCheck {
+}
