--- src/main/java/codesquad/handler/LogoutHandler.java
@@ -1,35 +1,0 @@
-package codesquad.handler;
-
-import java.time.Duration;
-
-import codesquad.annotation.RequestMapping;
-import codesquad.db.SessionDatabase;
-import codesquad.domain.Cookie;
-import codesquad.domain.HttpMethod;
-import codesquad.domain.HttpRequest;
-import codesquad.domain.HttpResponse;
-import codesquad.domain.HttpStatus;
-import codesquad.error.BaseException;
-import codesquad.utils.UserThreadLocal;
-
-public class LogoutHandler {
-
-	private LogoutHandler() {
-	}
-
-	@RequestMapping(httpMethod = HttpMethod.POST, url = "/logout")
-	public void doPost(HttpRequest request, HttpResponse response) {
-		if (!UserThreadLocal.isLogin()) {
-			throw new BaseException(HttpStatus.BAD_REQUEST, "You are not logged in");
-		}
-		Cookie sid = request.getCookie("SID");
-		SessionDatabase.getInstance().delete(sid.getValue());
-		removeCookie(response);
-		response.sendRedirect("/index.html");
-	}
-
-	private void removeCookie(HttpResponse response) {
-		Cookie cookie = new Cookie("SID", null, Duration.ofMillis(0), "/");
-		response.addHeader("Set-Cookie", cookie.toString());
-	}
-}
