+++ src/main/java/codesquad/handler/UserHandler.java
@@ -5,6 +5,7 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.UUID;
 
+import codesquad.annotation.LoginCheck;
 import codesquad.annotation.RequestMapping;
 import codesquad.db.SessionDatabase;
 import codesquad.db.UserDatabase;
@@ -16,14 +17,15 @@ import codesquad.domain.HttpResponse;
 import codesquad.domain.HttpStatus;
 import codesquad.domain.User;
 import codesquad.error.BaseException;
+import codesquad.utils.UserThreadLocal;
 
-public class LoginHandler {
+public class UserHandler {
 
-	private LoginHandler() {
+	private UserHandler() {
 	}
 
 	@RequestMapping(httpMethod = HttpMethod.POST, url = "/login")
-	public void doPost(HttpRequest request, HttpResponse response) {
+	public void login(HttpRequest request, HttpResponse response) {
 		Map<String, String> form = request.body().bodyToMap();
 		UserDatabase userDatabase = UserDatabase.getInstance();
 
@@ -39,12 +41,54 @@ public class LoginHandler {
 		setCookie(response, uuid);
 	}
 
+	@RequestMapping(httpMethod = HttpMethod.POST, url = "/logout")
+	@LoginCheck
+	public void logout(HttpRequest request, HttpResponse response) {
+		if (!UserThreadLocal.isLogin()) {
+			throw new BaseException(HttpStatus.BAD_REQUEST, "You are not logged in");
+		}
+		Cookie sid = request.getCookie("SID");
+		SessionDatabase.getInstance().delete(sid.getValue());
+		removeCookie(response);
+		response.sendRedirect("/index.html");
+	}
+
+	@RequestMapping(httpMethod = HttpMethod.POST, url = "/create")
+	public void signup(HttpRequest request, HttpResponse response) {
+		User user = getUser(request);
+		UserDatabase userDatabase = UserDatabase.getInstance();
+		userDatabase.insert(user.userId(), user);
+
+		setResponse(response);
+	}
+
+	@RequestMapping(httpMethod = HttpMethod.GET, url = "/user/list")
+	@LoginCheck
+	public void userList(HttpRequest request, HttpResponse response) {
+		response.sendRedirect("/user/list.html");
+	}
+
 	private void setResponse(HttpResponse response) {
 		response.setStatusLine(HttpStatus.FOUND);
 		HttpHeader header = new HttpHeader(new HashMap<>(Map.of("Location", "/index.html")));
 		header.setDefaultHeaders();
 		header.setHeaderValue("Content-Length", "0");
 		response.setHeader(header);
+	}
+
+	private User getUser(HttpRequest request) {
+		Map<String, String> params = request.body().bodyToMap();
+		String userId = params.get("userId");
+		String password = params.get("password");
+		String email = params.get("email");
+		String nickname = params.get("nickname");
+
+		return new User(userId, nickname, password, email);
+	}
+
+	private void removeCookie(HttpResponse response) {
+		Cookie cookie = new Cookie("SID", null, Duration.ofMillis(0), "/");
+		response.addHeader("Set-Cookie", cookie.toString());
 	}
 
 	private void setCookie(HttpResponse response, String uuid) {
