+++ src/main/resources/static/login_header.html
@@ -8,7 +8,7 @@
             <a class="btn btn btn_size_s" href="/post/new.html">글쓰기</a>
         </li>
         <li class="header__menu__item">
-            <a class="btn btn btn_size_s" href="/user/list.html">유저 목록</a>
+            <a class="btn btn btn_size_s" href="/user/list">유저 목록</a>
         </li>
         <li class="header__menu__item">
             <form action="/logout" method="POST" style="display:inline;">
