+++ src/main/resources/static/logout_header.html
@@ -5,7 +5,7 @@
       <a class="btn btn btn_size_s" href="/user/login.html">로그인</a>
     </li>
     <li class="header__menu__item">
-      <a class="btn btn btn_size_s" href="/user/list.html">유저 목록</a>
+      <a class="btn btn btn_size_s" href="/user/list">유저 목록</a>
     </li>
     <li class="header__menu__item">
       <a class="btn btn_contained btn_size_s" href="/register.html">회원 가입</a>
